ref: Rename to sentry-core and reexport
This moves the *src* of `sentry` to `sentry-core`, keeping the
`examples` and `tests` in place.

The `sentry` crate now just re-exports everything, and hands over
feature flags transparently.

The doctests have a `use sentry_core as sentry` in them to not require
code changes.
