+++ sentry-core/src/backtrace_support.rs
