+++ sentry-core/src/constants.rs
