+++ sentry-core/src/macros.rs
