+++ sentry-core/src/scope/mod.rs
