+++ sentry-core/src/scope/noop.rs
