+++ sentry-core/src/scope/real.rs
