+++ sentry-core/src/test.rs
@@ -9,6 +9,7 @@
 //! # Example usage
 //!
 //! ```
+//! # use sentry_core as sentry;
 //! use sentry::{capture_message, Level};
 //! use sentry::test::with_captured_events;
 //!
@@ -35,6 +36,7 @@ lazy_static::lazy_static! {
 /// Example usage:
 ///
 /// ```rust
+/// # use sentry_core as sentry;
 /// use std::sync::Arc;
 /// use sentry::{Hub, ClientOptions};
 /// use sentry::test::TestTransport;
