+++ sentry-core/src/transport.rs
