+++ sentry-core/src/utils.rs
