+++ sentry/Cargo.toml
@@ -10,7 +10,6 @@ documentation = "https://getsentry.github.io/sentry-rust"
 description = """
 Sentry (getsentry.com) client for rust ;)
 """
-build = "build.rs"
 edition = "2018"
 autoexamples = true
 
@@ -19,55 +18,35 @@ all-features = true
 
 [features]
 default = ["with_client_implementation", "with_default_transport", "with_panic", "with_failure", "with_device_info", "with_rust_info"]
-with_reqwest_transport = ["reqwest", "httpdate", "with_client_implementation"]
-with_curl_transport = ["curl", "httpdate", "serde_json", "with_client_implementation"]
-with_default_transport = ["with_reqwest_transport", "with_native_tls"]
-with_client_implementation = ["im", "url", "with_backtrace", "rand"]
-with_backtrace = ["backtrace", "regex"]
-with_panic = ["with_backtrace"]
-with_failure = ["failure", "with_backtrace"]
-with_log = ["log", "with_backtrace"]
-with_debug_to_log = ["log"]
-with_env_logger = ["with_log", "env_logger"]
-with_error_chain = ["error-chain", "with_backtrace"]
-with_device_info = ["libc", "hostname", "uname", "with_client_implementation"]
-with_rust_info = ["rustc_version", "with_client_implementation"]
-with_debug_meta = ["findshlibs", "with_client_implementation"]
-with_test_support = []
-with_rustls = ["reqwest/rustls-tls"]
-with_native_tls = ["reqwest/default-tls"]
+with_reqwest_transport = ["sentry-core/with_reqwest_transport"]
+with_curl_transport = ["sentry-core/with_curl_transport"]
+with_default_transport = ["sentry-core/with_default_transport"]
+with_client_implementation = ["sentry-core/with_client_implementation"]
+with_backtrace = ["sentry-core/with_backtrace"]
+with_panic = ["sentry-core/with_panic"]
+with_failure = ["sentry-core/with_failure"]
+with_log = ["sentry-core/with_log"]
+with_debug_to_log = ["sentry-core/with_debug_to_log"]
+with_env_logger = ["sentry-core/with_env_logger"]
+with_error_chain = ["sentry-core/with_error_chain"]
+with_device_info = ["sentry-core/with_device_info"]
+with_rust_info = ["sentry-core/with_rust_info"]
+with_debug_meta = ["sentry-core/with_debug_meta"]
+with_test_support = ["sentry-core/with_test_support"]
+with_rustls = ["sentry-core/with_rustls"]
+with_native_tls = ["sentry-core/with_native_tls"]
 
 [dependencies]
-backtrace = { version = "0.3.44", optional = true }
-url = { version = "2.1.1", optional = true }
-failure = { version = "0.1.6", optional = true }
-log = { version = "0.4.8", optional = true, features = ["std"] }
-sentry-types = { path = "../sentry-types", version = "0.15.0" }
-env_logger = { version = "0.7.1", optional = true }
-reqwest = { version = "0.10.1", optional = true, features = ["blocking", "json"], default-features = false }
-lazy_static = "1.4.0"
-regex = { version = "1.3.4", optional = true }
-error-chain = { version = "0.12.1", optional = true }
-im = { version = "14.2.0", optional = true }
-libc = { version = "0.2.66", optional = true }
-hostname = { version = "0.3.0", optional = true }
-findshlibs = { version = "0.5.0", optional = true }
-rand = { version = "0.7.3", optional = true }
-httpdate = { version = "0.3.2", optional = true }
-curl = { version = "0.4.25", optional = true }
-serde_json = { version = "1.0.48", optional = true }
-
-[target."cfg(not(windows))".dependencies]
-uname = { version = "0.1.1", optional = true }
-
-[build-dependencies]
-rustc_version = { version = "0.2.3", optional = true }
+sentry-core = { version = "0.18.0", path = "../sentry-core", default-features = false }
 
 [dev-dependencies]
 failure_derive = "0.1.6"
-pretty_env_logger = "0.4.0"
 actix-web = { version = "0.7.19", default-features = false }
-tokio = { version = "0.2", features = ["full"] }
+tokio = { version = "0.2", features = ["macros"] }
+log = { version = "0.4.8", features = ["std"] }
+failure = "0.1.6"
+pretty_env_logger = "0.4.0"
+error-chain = "0.12.1"
 
 [[example]]
 name = "error-chain-demo"
